Rename 'information' to 'announcement'.
