+++ app-android/build.gradle.kts
@@ -68,7 +68,7 @@ dependencies {
     implementation(projects.featureSessions)
     implementation(projects.featureContributors)
     implementation(projects.featureMap)
-    implementation(projects.featureInformation)
+    implementation(projects.featureAnnouncement)
     implementation(projects.featureAbout)
     implementation(projects.coreData)
     implementation(projects.coreDesignsystem)
