+++ app-android/src/main/res/values/strings.xml
@@ -2,7 +2,7 @@
     <!-- Navigation Drawer -->
     <string name="title_sessions">タイムライン</string>
     <string name="title_about">DroidKaigiとは</string>
-    <string name="title_information">お知らせ</string>
+    <string name="title_announcement">お知らせ</string>
     <string name="title_map">フロアマップ</string>
     <string name="title_sponsors">スポンサー</string>
     <string name="title_contributors">Contributor</string>
