+++ feature-announcement/src/main/AndroidManifest.xml
