+++ feature-announcement/src/main/java/io/github/droidkaigi/confsched2022/feature/announcement/AnnouncementNavGraph.kt
@@ -1,20 +1,20 @@
-package io.github.droidkaigi.confsched2022.feature.information
+package io.github.droidkaigi.confsched2022.feature.announcement
 
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
 
-fun NavGraphBuilder.informationGraph(
+fun NavGraphBuilder.announcementGraph(
     showNavigationIcon: Boolean,
     onNavigationIconClick: () -> Unit,
 ) {
-    composable(route = InformationNavGraph.informationRoute) {
-        InformationScreenRoot(
+    composable(route = AnnouncementNavGraph.announcementRoute) {
+        AnnouncementScreenRoot(
             showNavigationIcon = showNavigationIcon,
             onNavigationIconClick = onNavigationIconClick
         )
     }
 }
 
-object InformationNavGraph {
-    const val informationRoute = "information"
+object AnnouncementNavGraph {
+    const val announcementRoute = "announcement"
 }
