+++ feature-announcement/src/main/res/values-en/strings.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="announcement_top_app_bar_title">Announcement</string>
+</resources>
