+++ feature-announcement/src/test/resources/robolectric.properties
