--- feature-information/src/main/res/values/strings.xml
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <string name="information_top_app_bar_title">お知らせ</string>
-</resources>
